refactor: move metricNames to constnames
